Corrected path calculation, everything is working as expected

+++ src/main/java/com/ecse611/ipdetect/Weight.java
@@ -2,8 +2,8 @@ package com.ecse611.ipdetect;
 
 public class Weight {
 	public static final double SAME = 1;
-	public static final double CALLING = 3/14;
-	public static final double CALLED = 3/14;
-	public static final double DEF = 3/7;
-	public static final double THRESHOLD = 7/10;
+	public static final double CALLING = (double) 3/14;
+	public static final double CALLED = (double) 3/14;
+	public static final double DEF = (double) 3/7;
+	public static final double THRESHOLD = (double) 7/10;
 }
